add profile status card component
